backend/requirements.txt
rq
redis
flask==1.1.1
pycryptodome==3.9.4
requests==2.22.0
httpie

